% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_kmeans}
\alias{benchmark_kmeans}
\title{Benchmark My K-Means vs Base K-Means}
\usage{
benchmark_kmeans(data, centers, nstart = 10)
}
\arguments{
\item{data}{A numeric matrix or data frame.}

\item{centers}{The number of clusters.}

\item{nstart}{The number of random sets to be chosen.}
}
\value{
A benchmark result showing the performance comparison.
}
\description{
This function compares the performance of `my_kmeans` and base R's `kmeans`.
}
\examples{
data <- matrix(rnorm(1000), ncol = 2)
benchmark_kmeans(data, centers = 3, nstart = 10)
}
