% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{my_kmeans}
\alias{my_kmeans}
\title{K-Means Clustering Implementation}
\usage{
my_kmeans(data, centers, nstart = 10)
}
\arguments{
\item{data}{A numeric matrix or data frame.}

\item{centers}{The number of clusters.}

\item{nstart}{The number of random sets to be chosen.}
}
\value{
A list with cluster assignments, cluster centers, and within-cluster sum of squares.
}
\description{
This function performs K-Means clustering on a dataset.
}
\examples{
data <- matrix(rnorm(100), ncol = 2)
result <- my_kmeans(data, centers = 3, nstart = 10)
}
